requirements.txt
Pillow==3.3.0
pytesseract==0.1.6
